requirements.txt
asgiref==3.7.2
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
install==1.3.5
psycopg2-binary==2.9.6
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
